fix(bug): Missing icon on Linux (#29)
closes #23

The issue with the icon not displaying correctly on linux was because of an incorrect icon path.

I found this https://stackoverflow.com/questions/30845416/how-to-go-back-1-folder-level-wih-dirname that helped solve our issue 😃

The __dirname node variable gives the current path of the javascript file we are in. In this case the __dirname was giving the path inside /src which we were then concatenating the /icon.ico to it. Since there is no icon.ico in this folder the icon was not showing up.

I am using ubuntu and as you can see the beautiful icon is now displaying properly 🎉